[Fix] PluginBridge: Even though InterlockedSomething WinAPI functions work with uint32_t with the VS2010 provided Windows SDK, this is not documented. Older SDK versions do not provide unsigned overloads for C++ and thus compilation fails with VS2008. Replace uint32_t with the documented supported type for the InterlockedSomething functions: LONG.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@3914 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ pluginBridge/Bridge.h
@@ -16,7 +16,7 @@
 class PluginBridge : protected BridgeCommon
 {
 public:
-	static uint32_t instanceCount;
+	static LONG instanceCount;
 	static Event sigQuit;
 
 protected:
@@ -28,7 +28,7 @@ protected:
 	HINSTANCE library;
 	ERect windowSize;
 	HWND window, windowParent;
-	uint32_t isProcessing;
+	LONG isProcessing;
 
 	// Static memory for host-to-plugin pointers
 	union
